+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Version link in navbar not working.
 
 ### Changed
+- Images are now optimized before being proxied (70-80% smaller).
 - The Git hash is now included in versions that aren't built against a release tag.